Rename psi.inventory -> data_acquisition
- Moved out of psi because everything else in that directory is about
  the actual calculations
- Renamed because it deals with both inventory *and* waveform data
